move to netutils for config operations
