+++ nornir_nautobot/utils/helpers.py
@@ -1,0 +1,18 @@
+"""A set of helper utilities."""
+
+import errno
+import os
+import logging
+
+LOGGER = logging.getLogger(__name__)
+
+
+def make_folder(folder):
+    """Helper method to sanely create folders."""
+    if not os.path.exists(folder):
+        # Still try and except, since their may be race conditions.
+        try:
+            os.makedirs(folder)
+        except OSError as error:
+            if error.errno != errno.EEXIST:
+                raise
